Update ceph to 45c81859a5ff37c6cbbce0d260801e9f25354515
45c81859a5ff37c6cbbce0d260801e9f25354515 Merge "Add support for setting globally ms_bind_ipv6"
6cf8a4ea88ef031cab4b18c2a666badd779948db Merge "Adds public_addr setting."
3d7d647789455e36993fdd97c68408093cc2b16c Add support for setting globally ms_bind_ipv6
e26cb47661703e30553bd6fc6cb79f4f932419f4 drop mock dependency
5252808a9b6b5e883c92f1e42155a95d467ecbc9 Adds public_addr setting.
d387b0725f8eb01f2e9b572f7e3c22d5e00c67d9 Merge "Added default timeout for exec resources"
749a2a8026d4f5d3cc54c119b7be8dab12595eb9 Merge "fix mon creation to be more idempotent"
8fd97b5530c21ebfc82044fdeb368097cf72134a Merge "Be more specific with CentOS7 package exclude"
6a03609cce9f7b1f379230ef01d844723cf051da Added default timeout for exec resources
5096c97ddfbd4b491eb6050db41df04a93c0a8b9 fix mon creation to be more idempotent
618a5b2553590be76fe6a4c792b6c23d0a04bf0d Be more specific with CentOS7 package exclude

+++ ceph/manifests/mon.pp
@@ -51,6 +51,9 @@
 # [*keyring*] Path of the [mon.] keyring file
 #   Optional. $key and $keyring are mutually exclusive.
 #
+# [*exec_timeout*] The default exec resource timeout, in seconds
+#   Optional. Defaults to $::ceph::params::exec_timeout
+#
 define ceph::mon (
   $ensure = present,
   $public_addr = undef,
@@ -58,6 +61,7 @@ define ceph::mon (
   $authentication_type = 'cephx',
   $key = undef,
   $keyring  = undef,
+  $exec_timeout = $::ceph::params::exec_timeout,
   ) {
 
     # a puppet name translates into a ceph id, the meaning is different
@@ -133,6 +137,9 @@ define ceph::mon (
         command => "/bin/true # comment to satisfy puppet syntax requirements
 set -ex
 touch /etc/ceph/${cluster_name}.client.admin.keyring",
+        unless  => "/bin/true # comment to satisfy puppet syntax requirements
+set -ex
+test -e /etc/ceph/${cluster_name}.client.admin.keyring",
       }
       ->
       exec { $ceph_mkfs:
@@ -152,13 +159,18 @@ if [ ! -d \$mon_data ] ; then
   fi
 fi
 ",
+        unless    => "/bin/true # comment to satisfy puppet syntax requirements
+set -ex
+mon_data=\$(ceph-mon ${cluster_option} --id ${id} --show-config-value mon_data)
+test -d  \$mon_data
+",
         logoutput => true,
+        timeout   => $exec_timeout,
       }
       ->
       service { $mon_service:
         ensure => running,
       }
-
 
       if $authentication_type == 'cephx' {
         if $key {
@@ -166,6 +178,10 @@ fi
 
           exec { "rm-keyring-${id}":
             command => "/bin/rm ${keyring_path}",
+            unless  => "/bin/true # comment to satisfy puppet syntax requirements
+set -ex
+test ! -e ${keyring_path}
+",
           }
         }
       }
@@ -188,6 +204,7 @@ mon_data=\$(ceph-mon ${cluster_option} --id ${id} --show-config-value mon_data)
 test ! -d \$mon_data
 ",
         logoutput => true,
+        timeout   => $exec_timeout,
       } -> Package<| tag == 'ceph' |>
     }
   }
